% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSet.R
\name{basicSet}
\alias{basicSet}
\title{Basic Features}
\usage{
basicSet(text)
}
\arguments{
\item{text}{character A vector of texts, each of which will be tallied for DICE features.}
}
\description{
Simple features as inputs to the DICE model
}
