% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diceNGrams}
\alias{diceNGrams}
\title{Pre-trained advice concreteness features}
\format{
A (truncated) matrix of ngram feature counts for alignment to the pre-trained glmnet models.
}
\source{
Yeomans et al., (2024). A Natural Language Processing Model for Conflict Expression in Conversation
}
\usage{
diceNGrams
}
\description{
For internal use only. This dataset demonstrates the ngram features that are used for the pre-trained models.
}
\keyword{datasets}
