% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intenseModel}
\alias{intenseModel}
\title{Pre-trained Intensity Detection Model}
\format{
A pre-trained glmnet model
}
\source{
Yeomans et al. (2024). A Natural Language Processing Model for Conflict Expression in Conversation
}
\usage{
intenseModel
}
\description{
This model was pre-trained on wikipedia editor discussions hand-labelled for conflict intensity.
}
\keyword{datasets}
