% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polymodel}
\alias{polymodel}
\title{Polynomial pre-trained fit}
\format{
A pre-trained polynomial equation
}
\usage{
polymodel
}
\description{
This calculates the polynomial projection of the simple features used during model training
}
\keyword{datasets}
